Update Guacamole env and widen QEMU memory
Use GUAC_USER and GUAC_PASS for Guacamole credentials. Make tunnel_path,
api_path and connection_prefix local-only in build_env_config instead of
storing them in EnvConfig. Remove the delete_with_vnc_disable helper.
Change QemuConfig.memory_mb from u32 to u64.

+++ backend/src/qemu.rs
@@ -46,7 +46,7 @@ pub enum QemuError {
 #[derive(Debug, Clone)]
 pub struct QemuConfig {
     /// Memory size in MB
-    pub memory_mb: u32,
+    pub memory_mb: u64,
     /// Number of CPU cores
     pub cpu_cores: u32,
     /// Enable KVM acceleration
